chore: apply pr comments
