+++ src/resources/Enums.ts
@@ -1168,3 +1168,12 @@ export enum ProductsSortByType {
 export enum FeedbackConsumerType {
     SOURCE_MANAGEMENT_IMPROVEMENTS = 'SOURCE_MANAGEMENT_IMPROVEMENTS',
 }
+
+export enum ApiKeyStatus {
+    ENABLED='EXPIRED',
+    DISABLED='DISABLED',
+    DISABLED_FOR_INACTIVITY='DISABLED_FOR_INACTIVITY',
+    EXPIRED='EXPIRED',
+    SOON_TO_BE_EXPIRED='SOON_TO_BE_EXPIRED',
+    SOON_TO_BE_DISABLED='SOO_TO_BE_DISABLED'
+}
